Several occurences of real(8) etc converted to real(dp) etc. Routines and interfaces for fwftt, invfft moved from CPV/ to/ Modules/fft_interfaces.f90. This will be used soon everywhere in place of calls to cft3, cft3s
git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@7018 c92efa57-630b-4861-b058-cf58834340f0

+++ CPV/phasefactor.f90
@@ -258,8 +258,8 @@
       REAL(DP)    :: ainvb(3,3)
 ! local           
       integer :: i,j,k, is, ia, ig, isa
-      complex(8), allocatable:: ei1b(:,:), ei2b(:,:), ei3b(:,:)
-      real(8), allocatable :: taus(:,:)
+      complex(dp), allocatable:: ei1b(:,:), ei2b(:,:), ei3b(:,:)
+      real(dp), allocatable :: taus(:,:)
 !
       allocate(ei1b(-nr1b:nr1b,nat))
       allocate(ei2b(-nr2b:nr2b,nat))
